CMS 09/15 17:08

+++ src/main/webapp/WEB-INF/views/admin/passView.jsp
@@ -34,7 +34,7 @@
         <tr>
         	<td colspan="2">
         		<a href="/admin/passUpdateForm?pno=${view.pno}">수정</a> / 
-        		<a href="/admin/passDelete?pno=${view.pno}">삭제</a> /
+        		<a href="/admin/passDelete?pno=${view.pno}" onclick="return confirm('정말 삭제하시겠습니까?');">삭제</a> /
         		<a href="/admin/adminPassList">목록</a>  
         	</td>
         </tr>
